Loaders pass to the next stage, still more fixes are needed

+++ src/emulator/storage.rs
@@ -16,7 +16,7 @@ impl<
 {
     pub const CAPACITY: usize = CYLINDERS * HEADS * SECTORS_PER_TRACK * SECTOR_SIZE;
 
-    fn chs_to_lba(&self, c: u8, h: u8, s: u8) -> usize {
+    fn chs_to_lba(&self, c: u16, h: u8, s: u8) -> usize {
         let track_offset = (c as usize * HEADS + h as usize) * SECTORS_PER_TRACK;
         let sector_offset = s as usize - 1;
         (track_offset + sector_offset) * SECTOR_SIZE
@@ -35,7 +35,7 @@ impl<
     const SECTOR_SIZE: usize,
 > Floppy for GenericFloppy<CYLINDERS, HEADS, SECTORS_PER_TRACK, SECTOR_SIZE>
 {
-    fn read_chs_sectors(&self, c: u8, h: u8, s: u8, count: u8) -> &[u8] {
+    fn read_chs_sectors(&self, c: u16, h: u8, s: u8, count: u8) -> &[u8] {
         let start_lba = self.chs_to_lba(c, h, s);
         let start_byte = start_lba;
         let bytes_to_read = count as usize * SECTOR_SIZE;
@@ -44,7 +44,7 @@ impl<
 }
 
 pub trait Floppy {
-    fn read_chs_sectors(&self, c: u8, h: u8, s: u8, count: u8) -> &[u8];
+    fn read_chs_sectors(&self, c: u16, h: u8, s: u8, count: u8) -> &[u8];
 }
 
 pub type Floppy525_160 = GenericFloppy<40, 1, 8, 512>;
